Refactor Invoicing API and add invoice endpoints
This commit includes a number of changes to structure the Invoicing API more effectively and provide a clearer separation of concerns. It includes introducing configuration classes and restructuring the `Program.cs` file. It also involves the implementation of an Invoice service and the addition of new endpoints to the API for creating, retrieving single and all invoices. Furthermore, the API documentation was updated to reflect these new endpoints. Note that the deep copy extension for JSON serialization also now handles potential null returns.